+++ examples/dicom/deid-dicom-example.py
@@ -11,7 +11,7 @@ import os
 
 # This will get a set of example cookie dicoms
 base = get_dataset('dicom-cookies')
-dicom_files = get_files(base)
+dicom_files = list(get_files(base)) # todo : consider using generator functionality
 
 
 # This is the function to get identifiers